Правки
Два файла modal, ресет формы
