+++ js/modal-error.js
@@ -2,9 +2,9 @@ const GET_ERROR_TEXT = 'Ошибка при загрузке данных';
 const POST_ERROR_TEXT = 'Ошибка отправки объявления';
 
 const errorTemplate = document.querySelector('#error').content.querySelector('.error');
-const successTemplate = document.querySelector('#success').content.querySelector('.success');
+
 let errorClone;
-let successClone;
+
 
 const createError = (text) => {
   errorClone = errorTemplate.cloneNode(true);
@@ -14,13 +14,9 @@ const createError = (text) => {
 };
 
 const onDocumentKeydown = (evt) => {
-  evt.preventDefault();
   if (evt.key === 'Escape' && errorClone) {
     errorClone.remove();
-    document.removeEventListener('keydown', onDocumentKeydown);
-  }
-  if (evt.key === 'Escape' && successClone) {
-    successClone.remove();
+    evt.preventDefault();
     document.removeEventListener('keydown', onDocumentKeydown);
   }
 };
@@ -35,15 +31,4 @@ const renderPostErrorMessage = () => {
   document.addEventListener('keydown', onDocumentKeydown);
 };
 
-const createSuccess = () => {
-  successClone = successTemplate.cloneNode(true);
-  successClone.addEventListener('click', () => successClone.remove());
-  document.body.append(successClone);
-};
-
-const renderSuccessMessage = () => {
-  createSuccess();
-  document.addEventListener('keydown', onDocumentKeydown);
-};
-
-export { renderGetErrorMessage, renderPostErrorMessage, renderSuccessMessage };
+export { renderGetErrorMessage, renderPostErrorMessage };
